requirements.txt
supabase==0.5.8
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
streamlit==1.29.0
joblib==1.3.2

